[Tests] Add parallel run for structuredgridfactorytest  * Modified test to run in parallel  * Made parallel UG grid check universally usable
